Break the Elytron properties realms into their own modules
