+++ bom/deployment/pom.xml
@@ -348,6 +348,11 @@
             </dependency>
             <dependency>
                 <groupId>io.quarkus</groupId>
+                <artifactId>quarkus-elytron-security-properties-file-deployment</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.quarkus</groupId>
                 <artifactId>quarkus-security-deployment</artifactId>
                 <version>${project.version}</version>
             </dependency>
