+++ bom/runtime/pom.xml
@@ -314,6 +314,11 @@
             </dependency>
             <dependency>
                 <groupId>io.quarkus</groupId>
+                <artifactId>quarkus-elytron-security-properties-file</artifactId>
+                <version>${project.version}</version>
+            </dependency>
+            <dependency>
+                <groupId>io.quarkus</groupId>
                 <artifactId>quarkus-elytron-security-oauth2</artifactId>
                 <version>${project.version}</version>
             </dependency>
