+++ extensions/elytron-security-properties-file/deployment/src/test/java/io/quarkus/security/test/TestCharClone.java
