+++ extensions/elytron-security-properties-file/deployment/src/test/resources/application-custom-auth-embedded.properties
@@ -1,0 +1,14 @@
+quarkus.users.embedded.enabled=true
+quarkus.users.embedded.users.scott=jb0ss
+quarkus.users.embedded.users.stuart=test
+quarkus.users.embedded.users.jdoe=p4ssw0rd
+quarkus.users.embedded.users.noadmin=n0Adm1n
+quarkus.users.embedded.roles.scott=Admin,admin,Tester,user
+quarkus.users.embedded.roles.stuart=admin,user
+quarkus.users.embedded.roles.jdoe=NoRolesUser
+quarkus.users.embedded.roles.noadmin=user
+quarkus.users.embedded.auth-mechanism=CUSTOM
+
+#quarkus.log.min-level=DEBUG
+#quarkus.log.level=DEBUG
+#quarkus.log.console.level=DEBUG
