+++ extensions/elytron-security-properties-file/deployment/src/test/resources/application-custom-auth.properties
@@ -1,0 +1,8 @@
+quarkus.users.file.enabled=true
+quarkus.users.file.users=test-users.properties
+quarkus.users.file.roles=test-roles.properties
+quarkus.users.file.auth-mechanism=CUSTOM
+
+#quarkus.log.min-level=DEBUG
+#quarkus.log.level=DEBUG
+#quarkus.log.console.level=DEBUG
