+++ extensions/elytron-security-properties-file/deployment/src/test/resources/application.properties
@@ -1,0 +1,15 @@
+quarkus.users.file.enabled=true
+quarkus.users.file.users=test-users.properties
+quarkus.users.file.roles=test-roles.properties
+
+quarkus.log.level=DEBUG
+quarkus.log.console.enable=true
+quarkus.log.console.level=DEBUG
+quarkus.log.file.enable=true
+# Send output to a trace.log file under the /tmp directory
+quarkus.log.file.path=/tmp/trace.log
+quarkus.log.file.level=TRACE
+quarkus.log.file.format=%d{HH:mm:ss} %-5p [%c{2.}]] (%t) %s%e%n
+# Set 2 categories (io.quarkus.smallrye.jwt, io.undertow.request.security) to TRACE level
+quarkus.log.category."io.quarkus.arc".level=TRACE
+quarkus.log.category."io.undertow.request.security".level=TRACE
