+++ extensions/elytron-security-properties-file/deployment/src/test/resources/logging.properties
@@ -1,0 +1,18 @@
+loggers=org.jboss.logmanager
+
+# Root logger
+logger.level=INFO
+logger.handlers=CONSOLE
+
+logger.org.jboss.logmanager.useParentHandlers=true
+logger.org.jboss.logmanager.level=INFO
+
+handler.CONSOLE=org.jboss.logmanager.handlers.ConsoleHandler
+handler.CONSOLE.formatter=PATTERN
+handler.CONSOLE.properties=autoFlush,target
+handler.CONSOLE.autoFlush=true
+handler.CONSOLE.target=SYSTEM_OUT
+
+formatter.PATTERN=org.jboss.logmanager.formatters.PatternFormatter
+formatter.PATTERN.properties=pattern
+formatter.PATTERN.pattern=%d{HH:mm:ss,SSS} %-5p [%c] (%t) %s%E%n
