+++ extensions/elytron-security-properties-file/deployment/src/test/resources/test-roles.properties
@@ -1,0 +1,4 @@
+scott=Admin,admin,Tester,user
+jdoe=NoRolesUser
+stuart=admin,user
+noadmin=user
