+++ extensions/elytron-security-properties-file/deployment/src/test/resources/test-users.properties
@@ -1,0 +1,5 @@
+scott=jb0ss
+jdoe=p4ssw0rd
+stuart=test
+noadmin=n0Adm1n
+alice=alice
