+++ extensions/pom.xml
@@ -89,6 +89,7 @@
         <!-- Security -->
         <module>security</module>
         <module>elytron-security</module>
+        <module>elytron-security-properties-file</module>
         <module>elytron-security-oauth2</module>
         <module>smallrye-jwt</module>
         <module>vertx-keycloak</module>
