+++ extensions/security/deployment/src/main/java/io/quarkus/security/deployment/SecurityConfig.java
@@ -1,0 +1,20 @@
+package io.quarkus.security.deployment;
+
+import java.util.List;
+
+import io.quarkus.runtime.annotations.ConfigItem;
+import io.quarkus.runtime.annotations.ConfigPhase;
+import io.quarkus.runtime.annotations.ConfigRoot;
+
+/**
+ *
+ */
+@ConfigRoot(phase = ConfigPhase.BUILD_TIME)
+public final class SecurityConfig {
+
+    /**
+     * List of security providers to enable for reflection
+     */
+    @ConfigItem
+    public List<String> securityProviders;
+}
