+++ integration-tests/elytron-security/pom.xml
@@ -21,7 +21,7 @@
     <dependencies>
         <dependency>
             <groupId>io.quarkus</groupId>
-            <artifactId>quarkus-elytron-security</artifactId>
+            <artifactId>quarkus-elytron-security-properties-file</artifactId>
         </dependency>
 
         <!-- JAX-RS -->
