+++ integration-tests/elytron-security/src/main/resources/META-INF/microprofile-config.properties
@@ -4,6 +4,6 @@ quarkus.ssl.native=false
 
 quarkus.security.security-providers=SunRsaSign
 
-quarkus.security.embedded.enabled=true
-quarkus.security.embedded.users.scott=jb0ss
-quarkus.security.embedded.roles.scott=Admin,admin,Tester,user
+quarkus.users.embedded.enabled=true
+quarkus.users.embedded.users.scott=jb0ss
+quarkus.users.embedded.roles.scott=Admin,admin,Tester,user
