+++ integration-tests/main/src/main/resources/application.properties
@@ -15,7 +15,7 @@ quarkus.datasource.url: ${datasource.url}
 quarkus.datasource.driver: ${datasource.driver}
 quarkus.datasource.username: ${datasource.username}
 quarkus.datasource.password: ${datasource.password}
-quarkus.security.file.enabled=true
+quarkus.users.file.enabled=true
 quarkus.resteasy.gzip.enabled=true
 quarkus.resteasy.gzip.max-input=10
 quarkus.http.limits.max-body-size=1K
